Ecto v3.0.0-dev Ecto.Adapter.Storage behaviour

Specifies the adapter storage API.

Callbacks

storage_down(options)

storage_down(options :: Keyword.t) ::
  :ok |
  {:error, :already_down} |
  {:error, term}

Drops the storage given by options.

Returns :ok if it was dropped successfully.

Returns {:error, :already_down} if the storage has already been dropped or {:error, term} in case anything else goes wrong.

Examples

storage_down(username: postgres,
             database: 'ecto_test',
             hostname: 'localhost')

storage_up(options)

storage_up(options :: Keyword.t) ::
  :ok |
  {:error, :already_up} |
  {:error, term}

Creates the storage given by options.

Returns :ok if it was created successfully.

Returns {:error, :already_up} if the storage has already been created or {:error, term} in case anything else goes wrong.

Examples

storage_up(username: postgres,
           database: 'ecto_test',
           hostname: 'localhost')