Ecto v3.0.0-dev Ecto.Adapters.Postgres

Adapter module for PostgreSQL.

It uses postgrex for communicating to the database and a connection pool, such as poolboy.

Features

Full query support (including joins, preloads and associations)

Support for transactions

Support for data migrations

Support for ecto.create and ecto.drop operations

Support for transactional tests via Ecto.Adapters.SQL

Options

Postgres options split in different categories described below. All options can be given via the repository configuration:

config :your_app, YourApp.Repo,
  adapter: Ecto.Adapters.Postgres,
  ...

Non-compile time options can also be returned from the repository init/2 callback.

Compile time options

Those options should be set in the config file and require recompilation in order to make an effect.

:adapter - The adapter name, in this case, Ecto.Adapters.Postgres

Connection options

:pool - The connection pool module, defaults to DBConnection.Poolboy

:pool_timeout - The default timeout to use on pool calls, defaults to 5000

:timeout - The default timeout to use on queries, defaults to 15000

:hostname - Server hostname

:port - Server port (default: 5432)

:username - Username

:password - User password

:ssl - Set to true if ssl should be used (default: false)

:ssl_opts - A list of ssl options, see Erlang’s ssl docs

:parameters - Keyword list of connection parameters

:connect_timeout - The timeout for establishing new connections (default: 5000)

:prepare - How to prepare queries, either :named to use named queries or :unnamed to force unnamed queries (default: :named)

:socket_options - Specifies socket configuration

The :socket_options are particularly useful when configuring the size of both send and receive buffers. For example, when Ecto starts with a pool of 20 connections, the memory usage may quickly grow from 20MB to 50MB based on the operating system default values for TCP buffers. It is advised to stick with the operating system defaults but they can be tweaked if desired:

socket_options: [recbuf: 8192, sndbuf: 8192]

We also recommend developers to consult the Postgrex documentation for a complete listing of all supported options.

Storage options

:encoding - the database encoding (default: “UTF8”)

:template - the template to create the database from

:lc_collate - the collation order

:lc_ctype - the character classification

:dump_path - where to place dumped structures

Extensions

Both PostgreSQL and its adapter for Elixir, Postgrex, support an extension system. If you want to use custom extensions for Postgrex alongside Ecto, you must define a type module with your extensions. Create a new file anywhere in your application with the following:

Postgrex.Types.define(MyApp.PostgresTypes,
                      [MyExtension.Foo, MyExtensionBar] ++ Ecto.Adapters.Postgres.extensions(),
                      json: Poison)

Once your type module is defined, you can configure the repository to use it:

config :my_app, MyApp.Repo, types: MyApp.PostgresTypes

Functions

extensions()

All Ecto extensions for Postgrex.