Ecto v3.0.0-dev Ecto.Association.HasThrough

The association struct for has_one and has_many through associations.

Its fields are:

cardinality - The association cardinality

field - The name of the association field on the schema

owner - The schema where the association was defined

owner_key - The key on the owner schema used for the association

through - The through associations

relationship - The relationship to the specified schema, default :child

Functions

on_repo_change(map, _, _, _)

Performs the repository change on the association.

Receives the parent changeset, the current changesets and the repository action options. Must return the persisted struct (or nil) or the changeset error.

Callback implementation for c:Ecto.Association.on_repo_change/4.