Ecto v3.0.0-dev Ecto.Schema.Metadata

Stores metadata of a struct.

State

The state of the schema is stored in the :state field and allows following values:

:build - the stuct was constructed in memory and is not persisted to database yet;

:loaded - the struct was loaded from database and represents persisted data;

:deleted - the struct was deleted and no longer represents persisted data.

Source

The :source field is a tuple tracking the database source (table or collection) where the struct is or should be persisted. It is represented as a tuple consisting of two fields:

source - tracking the table or collection name.

prefix - tracking the ecto prefix of the table or collection.

Context

The :context field represents additional state some databases require for proper updates of data. It is not used by the built-in adapters of Ecto.Adapters.Postres and Ecto.Adapters.MySQL.

Types

context()

context() :: any

source()

source() :: {Ecto.Schema.prefix, Ecto.Schema.source}

state()

state() :: :built | :loaded | :deleted

t()

t() :: %Ecto.Schema.Metadata{context: context, source: source, state: state}