Ecto v3.0.0-dev mix ecto.load

Loads the current environment’s database structure for the given repository from a previously dumped structure file.

The repository must be set under :ecto_repos in the current app configuration or given via the -r option.

This task needs some shell utility to be present on the machine running the task.

Database	Utility needed

PostgreSQL	psql

MySQL	mysql

Example

mix ecto.load

Command line options

-r, --repo - the repo to load the structure info into

-d, --dump-path - the path of the dump file to load from

-f, --force - do not ask for confirmation

-q, --quiet - run the command quietly

Functions

run(args)

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.