Phoenix.HTML v2.11.0-dev Phoenix.HTML

Helpers for working with HTML strings and templates.

When used, it imports the given modules:

Phoenix.HTML - functions to handle HTML safety;

Phoenix.HTML.Tag - functions for generating HTML tags;

Phoenix.HTML.Form - functions for working with forms;

Phoenix.HTML.Link - functions for generating links and urls;

Phoenix.HTML.Format - functions for formatting text;

HTML Safe

One of the main responsibilities of this module is to provide convenience functions for escaping and marking HTML code as safe.

By default, data output in templates is not considered safe:

<%= "<hello>" %>

will be shown as:

&lt;hello&gt;

User data or data coming from the database is almost never considered safe. However, in some cases, you may want to tag it as safe and show its “raw” contents:

<%= raw "<hello>" %>

Keep in mind most helpers will automatically escape your data and return safe content:

<%= content_tag :p, "<hello>" %>

will properly output:

<p>&lt;hello&gt;</p>

Types

safe()

safe() :: {:safe, iodata()}

Guaranteed to be safe

unsafe()

unsafe() :: Phoenix.HTML.Safe.t()

May be safe or unsafe (i.e. it needs to be converted)

Functions

escape_javascript(data)

escape_javascript(binary() | safe()) :: String.t()

Escapes quotes (double and single), double backslashes and other.

This function is useful in JavaScript responses when there is a need to escape html rendered from other templates, like in the following:

$("#container").append("<%= escape_javascript(render("post.html", post: @post)) %>");

html_escape(safe)

html_escape(unsafe()) :: safe()

Escapes the HTML entities in the given term, returning iodata.

iex> html_escape("<hello>")
{:safe, ["&lt;", "hello", "&gt;" | ""]}

iex> html_escape('<hello>')
{:safe, ["&lt;", 104, 101, 108, 108, 111, "&gt;"]}

iex> html_escape(1)
{:safe, "1"}

iex> html_escape({:safe, "<hello>"})
{:safe, "<hello>"}

raw(value)

raw(iodata() | safe()) :: safe()

Marks the given content as raw.

This means any HTML code inside the given string won’t be escaped.

iex> raw("<hello>")
{:safe, "<hello>"}
iex> raw({:safe, "<hello>"})
{:safe, "<hello>"}
iex> raw(nil)
{:safe, ""}

safe_to_string(arg)

safe_to_string(safe()) :: String.t()

Converts a safe result into a string.

Fails if the result is not safe. In such cases, you can invoke html_escape/1 or raw/1 accordingly before.

sigil_E(expr, opts) (macro)

Provides ~E sigil with HTML safe EEx syntax inside source files.

Does not raise on attempts to interpolate with #{}, but rather shows those characters literally, so it should be preferred over ~e.

iex> ~E"""
...> Hello <%= "world" %>
...> """
{:safe, [[["" | "Hello "] | "world"] | "\n"]}

sigil_e(expr, opts) (macro)

Provides ~e sigil with HTML safe EEx syntax inside source files.

Raises on attempts to interpolate with #{}, so ~E should be preferred.

iex> ~e"""
...> Hello <%= "world" %>
...> """
{:safe, [[["" | "Hello "] | "world"] | "\n"]}