Phoenix.HTML v2.11.0-dev Phoenix.HTML.FormData protocol

Converts a data structure into a Phoenix.HTML.Form struct.

Types

t()

t() :: term()

Functions

input_type(data, form, field)

input_type(t(), Phoenix.HTML.Form.t(), atom()) :: atom() | nil

Receives the given field and returns its input type (:text_input, :select, etc). Returns nil if the type is unknown.

input_validations(data, form, field)

input_validations(t(), Phoenix.HTML.Form.t(), atom()) :: Keyword.t()

Returns the HTML5 validations that would apply to the given field.

input_value(data, form, field)

input_value(t(), Phoenix.HTML.Form.t(), atom()) :: term()

Returns the value for the given field.

to_form(data, options)

to_form(t(), Keyword.t()) :: Phoenix.HTML.Form.t()

Converts a data structure into a Phoenix.HTML.Form struct.

The options are the same options given to form_for/4. It can be used by implementations to configure their behaviour and it must be stored in the underlying struct, with any custom field removed.

to_form(data, form, field, options)

to_form(t(), Phoenix.HTML.Form.t(), atom(), Keyword.t()) :: Phoenix.HTML.Form.t()

Converts the field in the given form based on the data structure into a Phoenix.HTML.Form struct.

The options are the same options given to inputs_for/4. It can be used by implementations to configure their behaviour and it must be stored in the underlying struct, with any custom field removed.