Phoenix.HTML v2.11.0-dev Phoenix.HTML.Format

Helpers related to formatting text.

Functions

text_to_html(string, opts \\ [])

text_to_html(Phoenix.HTML.unsafe(), Keyword.t()) :: Phoenix.HTML.safe()

Returns text transformed into HTML using simple formatting rules.

Two or more consecutive newlines \n\n are considered as a paragraph and text between them is wrapped in <p> tags. One newline \n is considered as a linebreak and a <br> tag is inserted.

Examples

iex> text_to_html("Hello\n\nWorld") |> safe_to_string
"<p>Hello</p>\n<p>World</p>\n"

iex> text_to_html("Hello\nWorld") |> safe_to_string
"<p>Hello<br>\nWorld</p>\n"

iex> opts = [wrapper_tag: :div, attributes: [class: "p"]]
...> text_to_html("Hello\n\nWorld", opts) |> safe_to_string
"<div class=\"p\">Hello</div>\n<div class=\"p\">World</div>\n"

Options

:escape - if false does not html escape input (default: true)

:wrapper_tag - tag to wrap each paragraph (default: :p)

:attributes - html attributes of the wrapper tag (default: [])

:insert_brs - if true insert <br> for single line breaks (default: true)