Phoenix.HTML v2.11.0-dev Phoenix.HTML.Link

Conveniences for working with links and URLs in HTML.

Functions

button(opts, opts)

Generates a button that uses a regular HTML form to submit to the given URL.

Useful to ensure that links that change data are not triggered by search engines and other spidering software.

Examples

button("hello", to: "/world")
#=> <button class="button" data-csrf="csrf_token" data-method="post" data-to="/world">hello</button>

button("hello", to: "/world", method: "get", class: "btn")
#=> <button class="btn" data-method="get" data-to="/world">hello</button>

Options

:to - the page to link to. This option is required

:method - the method to use with the button. Defaults to :post.

All other options are forwarded to the underlying button input.

link(text, opts)

Generates a link to the given URL.

Examples

link("hello", to: "/world")
#=> <a href="/world">hello</a>

link("<hello>", to: "/world")
#=> <a href="/world">&lt;hello&gt;</a>

link("<hello>", to: "/world", class: "btn")
#=> <a class="btn" href="/world">&lt;hello&gt;</a>

link("delete", to: "/the_world", data: [confirm: "Really?"])
#=> <a data-confirm="Really?" href="/the_world">delete</a>

# If you supply a method other than `:get`:
link("delete", to: "/everything", method: :delete)
#=> <a href="#" data-csrf="csrf_token" data-method="delete" data-to="/everything">delete</a>

# You can use a `do ... end` block too:
link to: "/hello" do
  "world"
end

Options

:to - the page to link to. This option is required

:method - the method to use with the link. In case the method is not :get, the link is generated inside the form which sets the proper information. In order to submit the form, JavaScript must be enabled

All other options are forwarded to the underlying <a> tag.

Data attributes

Data attributes are added as a keyword list passed to the data key. The following data attributes are supported:

data-confirm - shows a confirmation prompt before generating and submitting the form when :method is not :get.

JavaScript dependency

In order to support links where :method is not :get or use the above data attributes, Phoenix.HTML relies on JavaScript. You can load priv/static/phoenix_html.js into your build tool.

CSRF Protection

By default, CSRF tokens are generated through Plug.CSRFProtection. You can customize the CSRF token generation by configuring your own MFA:

config :phoenix_html, csrf_token_generator: {MyGenerator, :get_token, []}