Phoenix v1.4.0-dev Phoenix.Logger

Instrumenter to handle logging of various instrumentation events.

Parameter filtering

When logging parameters, Phoenix can filter out sensitive parameters such as passwords and tokens. Parameters to be filtered can be added via the :filter_parameters option:

config :phoenix, :filter_parameters, ["password", "secret"]

With the configuration above, Phoenix will filter any parameter that contains the terms password or secret. The match is case sensitive.

Phoenix’s default is ["password"].

Phoenix can filter all parameters by default and selectively keep parameters. This can be configured like so:

config :phoenix, :filter_parameters, {:keep, ["id", "order"]}

With the configuration above, Phoenix will filter all parameters, except those that match exactly id or order. If a kept parameter matches, all parameters nested under that one will also be kept.

Functions

phoenix_channel_join(arg1, compile, arg3)

phoenix_channel_receive(arg1, compile, meta)

phoenix_controller_call(arg1, arg2, arg3)