Plug v1.4.3 Plug.Conn.Cookies

Conveniences for encoding and decoding cookies.

Functions

decode(cookie)

Decodes the given cookies as given in a request header.

If a cookie is invalid, it is automatically discarded from the result.

Examples

iex> decode("key1=value1, key2=value2")
%{"key1" => "value1", "key2" => "value2"}

encode(key, opts \\ %{})

Encodes the given cookies as expected in a response header.