Plug v1.4.3 Plug.Crypto

Namespace and module for Crypto functionality.

Please see Plug.Crypto.KeyGenerator, Plug.Crypto.MessageEncryptor, and Plug.Crypto.MessageVerifier for more functionality.

Functions

mask(left, right)

Masks the token on the left with the token on the right.

Both tokens are required to have the same size.

masked_compare(left, right, mask)

Compares the two binaries (one being masked) in constant-time to avoid timing attacks.

It is assumed the right token is masked according to the given mask.

safe_binary_to_term(binary, opts \\ [])

A restricted version a :erlang.binary_to_term/1 that forbids possibly unsafe terms.

secure_compare(left, right)

Compares the two binaries in constant-time to avoid timing attacks.

See: http://codahale.com/a-lesson-in-timing-attacks/