Plug v1.4.3 Plug.Logger

A plug for logging basic request information in the format:

GET /index.html
Sent 200 in 572ms

To use it, just plug it into the desired module.

plug Plug.Logger, log: :debug

Options

:log - The log level at which this plug should log its request info. Default is :info.

Functions

call(conn, level)

Callback implementation for Plug.call/2.

init(opts)

Callback implementation for Plug.init/1.