Plug v1.4.3 Plug.Session.Store behaviour

Specification for session stores.

Types

cookie()

cookie() :: binary

The cookie value that will be sent in cookie headers. This value should be base64 encoded to avoid security issues.

session()

session() :: map

The session contents, the final data to be stored after it has been built with Plug.Conn.put_session/3 and the other session manipulating functions.

sid()

sid() :: term | nil

The internal reference to the session in the store.

Callbacks

delete(arg0, sid, arg2)

delete(Plug.Conn.t, sid, Plug.opts) :: :ok

Removes the session associated with given session id from the store.

get(arg0, cookie, arg2)

get(Plug.Conn.t, cookie, Plug.opts) :: {sid, session}

Parses the given cookie.

Returns a session id and the session contents. The session id is any value that can be used to identify the session by the store.

The session id may be nil in case the cookie does not identify any value in the store. The session contents must be a map.

init(arg0)

init(Plug.opts) :: Plug.opts

Initializes the store.

The options returned from this function will be given to get/3, put/4 and delete/3.

put(arg0, sid, any, arg3)

put(Plug.Conn.t, sid, any, Plug.opts) :: cookie

Stores the session associated with given session id.

If nil is given as id, a new session id should be generated and returned.