spotify_ex v2.0.7 Responder behaviour

Recieves http responses from the client and handles them accordingly.

Spotify API modules (Playlist, Album, etc) use Responder. When a request is made they giv the endpoint URL to the Client, which makes the request, and pass the response to handle_response. Each API module must build appropriate responses, so they add Responder as a behaviour, and implement the build_response/1 function.

Callbacks

build_response(map)

build_response(map()) :: any()