spotify_ex v2.0.7 Spotify.Authorization

Authorizes your app with Spotify

Spotify needs to verify your client id, and that your redirect uri matches what you set in your app settings (in the Spotify App dashboard). This is an external call, url provided by the url function.

Functions

url()

If you specified scopes in your config, uses scoped auth. Otherwise, unscoped. Use this function to make the redirect to Spotify for authorization.

## Example:

 defmodule OAuthController do
   # ...

   def authorize do
     redirect conn, external: Spotify.Authorization.url
   end
 end