spotify_ex v2.0.7 Spotify.Profile

Endpoints for retrieving information about a user’s profile.

There are two functions for each endpoint, one that actually makes the request, and one that provides the endpoint:

  Spotify.Playist.create_playlist(conn, "foo", "bar") # makes the POST request.
  Spotify.Playist.create_playlist_url("foo", "bar") # provides the url for the request.

https://developer.spotify.com/web-api/user-profile-endpoints/

Functions

build_response(body)

Implements the hook expected by the Responder behaviour

handle_response(arg)

me(conn)

Get detailed profile information about the current user (including the current user’s username). Spotify Documentation

Method: GET

Uses your auth token to find your profile.

Spotify.Profile.me(conn)
# => { :ok, %Spotify.Profile{..} }

me_url()

Get detailed profile information about the current user (including the current user’s username).

iex> Spotify.Profile.me_url
"https://api.spotify.com/v1/me"

user(conn, user_id)

Get public profile information about a Spotify user. Spotify Documentation

Method: GET

Spotify.Profile.user(conn, "123")
# => { :ok, %Spotify.Profile{..} }

user_url(user_id)

Get public profile information about a Spotify user.

iex> Spotify.Profile.user_url("123")
"https://api.spotify.com/v1/users/123"