spotify_ex v2.0.7 Spotify.Recommendation

Spotify can make recommendations for you by providing seed data. The recommendation comes with tracks and the seed object

*Note: The possibilities here are quite large. Please read the spotify documentation.

https://developer.spotify.com/web-api/get-recommendations/

Functions

build_response(body)

Implements the hook required by the Responder behavior.

get_recommendations(conn, params)

Create a playlist-style listening experience based on seed artists, tracks and genres.

The response generated by this can vary. Your milage may vary. Because of this, you may want to use get_recommendations_url and use your own implementation for this function.

Spotify Documentation

Method: GET

Params: The params for this endpoint are complex. Refer to spotify docs

Spotify.Recommendation.get_recommendations(seed_artists: "1,2" energy: "6")
# => { :ok, %Recommendation{tracks: tracks, seeds: seeds} }

get_recommendations_url(params)

Create a playlist-style listening experience based on seed artists, tracks and genres.

iex> Spotify.Recommendation.get_recommendations_url(limit: 5)
"https://api.spotify.com/v1/recommendations?limit=5"

handle_response(arg)